Removed inline styling and renamed some files
