+++ AngularTraining/user-management-staticdata/src/app/Activate/ActivateUser/app-activateuser.component.html
@@ -1,11 +1,8 @@
 <div class="box">
     <div class="card border-danger  mb-3" *ngFor="let user of (userList | deletedUsers)">
-
         <div class="card-body">
             <h5 class="card-title">{{user.name}}</h5>
-            <button type="button" class="btn btn-success" (click)="activateUser(user.id)"
-                style="float: right;">Activate</button>
+            <button type="button" class="btn btn-success btndisplay" (click)="activateUser(user.id)">Activate</button>
         </div>
     </div>
 </div>
-
