+++ AngularTraining/user-management-staticdata/src/app/Activate/ActivateUser/app-activateuser.component.ts
