+++ AngularTraining/user-management-staticdata/src/app/app-routing.module.ts
@@ -1,16 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { ActivateUsersComponent } from './Activate/ActivateUser/ActivateUser.component';
-import { DeleteUserComponent } from './feature2/DeleteUsers/DeleteUser.component';
+import { ActivateUsersComponent } from './Activate/ActivateUser/app-activateuser.component';
+import { DeleteUserComponent } from './feature2/DeleteUsers/app-deleteuser.component';
 import { LoginComponent } from './login/login.component';
 
 const appRoutes: Routes = [
     { path: '', component: LoginComponent },
-    { path: 'feature1', component: ActivateUsersComponent },
-    { path: 'feature2', component: DeleteUserComponent },
+    { path: 'activateuser', component: ActivateUsersComponent },
+    { path: 'deleteuser', component: DeleteUserComponent },
     {
-        path: 'feature3', 
+        path: 'userlisting', 
         loadChildren: ()=> import('./feature3/UserListingModule.module').then(
             (file) => file.UserListingModule
         )
