+++ AngularTraining/user-management-staticdata/src/app/app.component.css
@@ -17,6 +17,13 @@ div {
     font-weight: 600;
 }
 
+.ruler{
+    width: 90%;
+    margin-top: 5px;
+    margin-bottom: 5px;
+    border: none;
+    border-top: 1px solid gainsboro;
+}
 .nav-bar, .login-name {
     display: flex;
     justify-content: flex-end;
