+++ AngularTraining/user-management-staticdata/src/app/app.component.html
@@ -1,12 +1,12 @@
 <div class="login-name">{{loggedinUser}}</div>
 <div class="nav-bar">
   <div><a routerLink="/">Home</a></div>
-  <div routerLinkActive="active"><a routerLink="feature1">Deleted</a></div>
-  <div routerLinkActive="active"><a routerLink="feature2">Active</a></div>
-  <div routerLinkActive="active"><a routerLink="feature3">Manage</a></div>
+  <div routerLinkActive="active"><a routerLink="activateuser">Deleted</a></div>
+  <div routerLinkActive="active"><a routerLink="deleteuser">Active</a></div>
+  <div routerLinkActive="active"><a routerLink="userlisting">Manage</a></div>
 </div>
 
-<hr style="width: 90%; margin-top: 5px; margin-bottom: 5px; border: none;border-top: 1px solid gainsboro;" />
+<hr class="ruler"/>
 
 <router-outlet></router-outlet>
