+++ AngularTraining/user-management-staticdata/src/app/feature2/DeleteUsers/app-deleteuser.component.css
@@ -8,7 +8,9 @@
     font-family: calibri;
     font-size: large;
 }
-
+.btndisplay{
+    float: right;
+}
 div {
     margin-bottom: 5px;
 }
