+++ AngularTraining/user-management-staticdata/src/app/feature2/DeleteUsers/app-deleteuser.component.html
@@ -1,12 +1,9 @@
 <div class="box">
-
     <div class="card border-success mb-3" *ngFor="let user of (userList | activeUsers)">
-
         <div class="card-body">
             <h5 class="card-title">{{user.name}}</h5>
-            <button type="button" class="btn btn-danger" style="float: right;"
+            <button type="button" class="btn btn-danger btndisplay"
                 (click)="deactivateUser(user.id)">Deactivate</button>
         </div>
     </div>
 </div>
-
