+++ AngularTraining/user-management-staticdata/src/app/feature2/DeleteUsers/app-deleteuser.component.ts
@@ -4,7 +4,7 @@ import { UsersService } from 'src/app/users.service';
 import { Router } from '@angular/router';
 
 @Component({
-  selector: 'app-f2c1',
+  selector: 'app-deleteuser',
   templateUrl: './f2c1.component.html',
   styleUrls: ['./f2c1.component.css']
 })
@@ -12,8 +12,10 @@ export class DeleteUserComponent implements OnInit {
   userList = [];
   constructor(private usersService: UsersService, private route: Router) { }
   deactivateUser(userId) {
+    
     this.usersService.users.forEach(element => {
       if (userId === element.id) {
+        console.log(element.name);
         element.isDeleted = false;
       }
     });
