+++ AngularTraining/user-management-staticdata/src/app/feature3/ListUsers/app-listusers.component.css
@@ -5,7 +5,12 @@
     margin-left: 10%;
     
     }
-
+.linedisplay{
+    line-height: 50%;
+}
+.btndisplay{
+    float: right;
+}
 .box {
         border: 1px solid #90def8;
         border-radius: 20px;
