+++ AngularTraining/user-management-staticdata/src/app/feature3/UserInfo/app-userinfo.component.css
@@ -5,7 +5,10 @@
  div {
      margin-bottom: 10px;
  }
-
+.carddisplay{
+    max-width: 18rem;
+    margin-right: 10%
+}
  .activeclass{
     content:"Active";
     color: green;
