+++ AngularTraining/user-management-staticdata/src/app/feature3/UserInfo/app-userinfo.component.html
@@ -1,4 +1,4 @@
-<div class="card border-primary mb-3" style="max-width: 18rem;margin-right: 10%">
+<div class="card border-primary mb-3 carddisplay">
   <div class="card-header">{{user?.name}}</div>
   <div class="card-body text-primary">
     <h5 class="card-title activeclass" *ngIf="user?.isDeleted">Active</h5>
