+++ AngularTraining/user-management-staticdata/src/app/feature3/UserInfo/app-userinfo.component.ts
